+++ NEWS.md
@@ -1,5 +1,8 @@
 # bs4Dash 0.2.0
 - bs4Dash now works on shinyapp.io (previously, there was a problem with tab navigation)
+- close and collapse buttons not shown when the title is NULL in `bs4Card()`
+- `bs4ValueBox()` footer is not shown when the external link is NULL
+- `bs4ValueBox()` external link opens in a new page
 - add sidebar_collapsed argument to `bs4DashPage()` to control the sidebar opening at start
 - update `bs4DashGallery()`. Add a "theme" argument
 - add `ionicon()` (implementation of ionicons)